Extract per-endpoint route writing into RoutesGenerator class
